split file, remove unused code
